bugfix and raw testfile

+++ rawAnimPlayer.py
@@ -18,7 +18,7 @@ with open(sys.argv[1],"rb") as file:
 	animation = bytearray(file.read())
 
 frame=0;
-frames = len(animation)/(3*num_pixels)
+frames = (int)(len(animation)/(3*num_pixels))
 while(1):
 	neopixel_write.neopixel_write(pin,animation[3*num_pixels*frame:3*num_pixels*(frame+1)])
 	frame=(frame+1)%frames
